app/requirements.txt
Flask
flask-jwt-extended
Flask-SocketIO>=5.3.6
flask-bcrypt
flask-cors
Flask-Caching>=2.3
Flask-Session>=0.5.0
Flask-Limiter>=3.7
Flask-Migrate
requests
urllib3
python-dotenv
flask-sqlalchemy
SQLAlchemy
mysqlclient
pyrad
hypercorn
psycopg2-binary
eventlet>=0.36 
gunicorn
Werkzeug
itsdangerous
Jinja2
MarkupSafe
click
blinker
pycparser
cffi
pyOpenSSL
cryptography
six
PyJWT
routeros-api
flask_caching
redis
python-dateutil
routeros_api
redis
pymysql
flask-socketio>=5.3
python-socketio>=5.11
eventlet>=0.35
redis>=5.0
celery>=5.3
pymongo
